+++ resources/js/app.js
@@ -25,16 +25,20 @@ require('./components/searchbar')
 require('./components/clipboard-button')
 
 // Vue components
-window.Vue = require('vue')
+window.Vue = require('vue').default
 
 Vue.component('filter-sort-by', require('./components/filter/SortBy.vue').default);
 Vue.component('filter-checkbox', require('./components/filter/Checkbox.vue').default);
 Vue.component('filter-custom-select', require('./components/filter/CustomSelect.vue').default);
+Vue.component('user-collections-link', require('./components/user-collections/Link.vue').default);
 Vue.component('user-collections-nav-link', require('./components/user-collections/NavLink.vue').default);
 Vue.component('user-collections-favourite-button', require('./components/user-collections/FavouriteButton.vue').default);
 Vue.component('user-collections-clear-button', require('./components/user-collections/ClearButton.vue').default);
+Vue.component('user-collections-share-form', require('./components/user-collections/ShareForm.vue').default);
 Vue.component('color-widget', require('./components/ColorWidget.vue').default);
 Vue.component('year-slider', require('./components/YearSlider.vue').default);
+Vue.component('inline-input', require('./components/InlineInput.vue').default);
+Vue.component('copy-to-clipboard', require('./components/CopyToClipboard.vue').default);
 
 Vue.component('slider', require('./components/vue/slider').default);
 Vue.component('color-slider', require('./components/vue/color-slider').default);