+++ resources/js/components/user-collections/ShareForm.vue
@@ -1,0 +1,40 @@
+<template>
+    <form ref="form" :action="action" method="POST" v-on:submit.prevent="onSubmit">
+        <slot :editing="editing" :setEditing="setEditing"></slot>
+    </form>
+</template>
+
+<script>
+export default {
+    props: {
+        action: String,
+        creating: Boolean
+    },
+    data() {
+        return {
+            editing: false,
+        }
+    },
+    methods: {
+        setEditing(editing) {
+            this.editing = editing
+        },
+        onSubmit() {
+            const form = this.$refs.form
+
+            // Do a full reload submission on create
+            if (this.creating) {
+                form.submit()
+                return
+            }
+
+            fetch(this.action, {
+                method: 'POST',
+                body: new FormData(form),
+            })
+
+            this.editing = false
+        }
+    }
+}
+</script>