+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+
+## [2.32.0] - 2021-10-04
+### Added
+- paging for front page, articles and edu articles
+- user collection sharing
+
+### Changed
+- Laravel Mix & assets
+
 ## [2.31.0] - 2021-09-09
 ### Added
 - Scena.AI snippet